+++ jfleet-core/src/test/java/org/jfleet/postgres/PostgresTransactionPolicyTest.java
@@ -39,7 +39,7 @@ import org.slf4j.LoggerFactory;
 public class PostgresTransactionPolicyTest {
 
     private static Logger logger = LoggerFactory.getLogger(PostgresTransactionPolicyTest.class);
-    private static final long VERY_LOW_SIZE_TO_FREQUENT_LOAD_DATA = 10;
+    private static final int VERY_LOW_SIZE_TO_FREQUENT_LOAD_DATA = 10;
 
     private Supplier<Connection> provider;
